move cell line id column out of shared columns
